+++ ontask/scheduler/forms.py
@@ -35,6 +35,8 @@ class ScheduleBasicForm(FormWithPayload, forms.ModelForm):
         self.set_fields_from_dict(['name', 'description_text'])
         self.fields['execute'].initial = parse_datetime(
                 self._FormWithPayload__form_info.get('execute', ''))
+        self.fields['execute_until'].initial = parse_datetime(
+                self._FormWithPayload__form_info.get('execute_until', ''))
 
     def clean(self):
         """Verify that the date is corre    ct."""
@@ -43,15 +45,28 @@ class ScheduleBasicForm(FormWithPayload, forms.ModelForm):
         self.store_fields_in_dict([
             ('name', None),
             ('description_text', None),
-            ('execute', str(form_data['execute']))])
+            ('execute', str(form_data['execute'])),
+            ('execute_until', str(form_data['execute_until']))])
 
         # The executed time must be in the future
         now = datetime.datetime.now(pytz.timezone(settings.TIME_ZONE))
         when_data = self.cleaned_data.get('execute')
-        if when_data and when_data <= now:
+        until_data = self.cleaned_data.get('execute_until')
+        if when_data and not until_data and when_data <= now:
             self.add_error(
                 'execute',
                 _('Date/time must be in the future'))
+
+        if until_data:
+            if until_data <= when_data:
+                self.add_error(
+                    'execute_until',
+                    _('Final execution date/time must be after '
+                      + 'the previous one'))
+            elif until_data <= now:
+                self.add_error(
+                    'execute_until',
+                    _('Final execution date/time must be in the future'))
 
         return form_data
 
@@ -59,10 +74,11 @@ class ScheduleBasicForm(FormWithPayload, forms.ModelForm):
         """Define model, fields and widgets."""
 
         model = ScheduledAction
-        fields = ('name', 'description_text', 'execute')
+        fields = ('name', 'description_text', 'execute', 'execute_until')
         widgets = {
             'execute': DateTimePickerInput(options=date_time_widget_options),
-        }
+            'execute_until':
+                DateTimePickerInput(options=date_time_widget_options)}
 
 
 class ScheduleMailSubjectForm(FormWithPayload):
@@ -199,14 +215,18 @@ class ScheduleItemsForm(ScheduleBasicForm):
 
         if self.instance and not form_data['confirm_items']:
             self.instance.exclude_values = []
-            self.instance.save()
 
         return form_data
 
     class Meta(ScheduleBasicForm.Meta):
         """Define model, fields and widgets."""
 
-        fields = ('name', 'description_text', 'item_column', 'execute')
+        fields = (
+            'name',
+            'description_text',
+            'item_column',
+            'execute',
+            'execute_until')
 
 
 class ScheduleTokenForm(FormWithPayload):
@@ -280,6 +300,7 @@ class EmailScheduleForm(
             'name',
             'description_text',
             'execute',
+            'execute_until',
             'item_column',
             'subject',
             'cc_email',
@@ -317,6 +338,7 @@ class SendListScheduleForm(
             'name',
             'description_text',
             'execute',
+            'execute_until',
             'email_to',
             'subject',
             'cc_email',
@@ -346,6 +368,7 @@ class JSONScheduleForm(ScheduleTokenForm, ScheduleItemsForm):
             'name',
             'description_text',
             'execute',
+            'execute_until',
             'item_column',
             'confirm_items',
             'token')
@@ -361,6 +384,7 @@ class JSONListScheduleForm(ScheduleTokenForm, ScheduleBasicForm):
             'name',
             'description_text',
             'execute',
+            'execute_until',
             'token')
 
 
@@ -384,4 +408,5 @@ class CanvasEmailScheduleForm(ScheduleMailSubjectForm, ScheduleItemsForm):
             'item_column',
             'confirm_items',
             'subject',
-            'execute')
+            'execute',
+            'execute_until')