+++ ontask/tasks/tests/test_scheduled_actions.py
@@ -4,17 +4,19 @@
 
 import os
 import test
-from datetime import datetime
+from datetime import datetime, timedelta
+from psycopg2 import sql
 
 import pytz
 from celery.contrib.testing.worker import start_worker
 from django.conf import settings
 from django.contrib.auth import get_user_model
 from django.core import mail
+from django.db import connection
 
 from ontask import OnTaskSharedState, tasks
 from ontask.core.celery import app
-from ontask.models import Action, ScheduledAction
+from ontask.models import Action, ScheduledAction, Workflow
 
 
 class ScheduledActionTaskTestCase(test.OnTaskTestCase):
@@ -53,7 +55,8 @@ class ScheduledActionTaskTestCase(test.OnTaskTestCase):
             user=user,
             name='send email action',
             action=action,
-            execute=datetime.now(pytz.timezone(settings.TIME_ZONE)),
+            execute=datetime.now(pytz.timezone(settings.TIME_ZONE)).replace(
+                second=0),
             status=ScheduledAction.STATUS_PENDING,
             item_column=action.workflow.columns.get(name='email'),
             payload={
@@ -91,7 +94,8 @@ class ScheduledActionTaskTestCase(test.OnTaskTestCase):
             user=user,
             name='JSON scheduled action',
             action=action,
-            execute=datetime.now(pytz.timezone(settings.TIME_ZONE)),
+            execute=datetime.now(pytz.timezone(settings.TIME_ZONE)).replace(
+                second=0),
             status=ScheduledAction.STATUS_PENDING,
             item_column=action.workflow.columns.get(name='email'),
             payload={'token': token})
@@ -121,7 +125,8 @@ class ScheduledActionTaskTestCase(test.OnTaskTestCase):
             user=user,
             name='send list scheduled action',
             action=action,
-            execute=datetime.now(pytz.timezone(settings.TIME_ZONE)),
+            execute=datetime.now(pytz.timezone(settings.TIME_ZONE)).replace(
+                second=0),
             status=ScheduledAction.STATUS_PENDING,
             payload={
                 'email_to': 'recipient@bogus.com',
@@ -156,7 +161,8 @@ class ScheduledActionTaskTestCase(test.OnTaskTestCase):
             user=user,
             name='JSON List scheduled action',
             action=action,
-            execute=datetime.now(pytz.timezone(settings.TIME_ZONE)),
+            execute=datetime.now(pytz.timezone(settings.TIME_ZONE)).replace(
+                second=0),
             status=ScheduledAction.STATUS_PENDING,
             payload={'token': token})
         scheduled_item.save()
@@ -169,3 +175,115 @@ class ScheduledActionTaskTestCase(test.OnTaskTestCase):
         assert scheduled_item.status == ScheduledAction.STATUS_DONE
         assert len(json_outbox) == 1
         assert all(token in item['auth'] for item in json_outbox)
+
+    def test_scheduled_incremental_email_action(self):
+        """Test an incremental scheduled action."""
+        # Modify the data table so that initially all records have registered
+        # equal to alse
+        workflow = Workflow.objects.all().first()
+        with connection.cursor() as cursor:
+            query = sql.SQL('UPDATE {0} SET {1} = false').format(
+                sql.Identifier(workflow.get_data_frame_table_name()),
+                sql.Identifier('registered'))
+            cursor.execute(query)
+
+        user = get_user_model().objects.get(email='instructor01@bogus.com')
+
+        # User must exist
+        self.assertIsNotNone(user, 'User instructor01@bogus.com not found')
+        action = Action.objects.get(name='send email incrementally')
+
+        now = datetime.now(pytz.timezone(settings.TIME_ZONE)).replace(second=0)
+        scheduled_item = ScheduledAction(
+            user=user,
+            name='send email action incrementally',
+            action=action,
+            execute=now,
+            execute_until=now + timedelta(hours=1),
+            status=ScheduledAction.STATUS_PENDING,
+            item_column=action.workflow.columns.get(name='email'),
+            payload={
+                'subject': 'Email subject',
+                'cc_email': '',
+                'bcc_email': '',
+                'send_confirmation': False,
+                'track_read': False})
+        scheduled_item.save()
+
+        # Execute the scheduler for the first time
+        tasks.execute_scheduled_actions_task(True)
+
+        # Event stil pending, with no values in exclude values
+        scheduled_item.refresh_from_db()
+        assert scheduled_item.status == ScheduledAction.STATUS_PENDING
+        assert scheduled_item.exclude_values == []
+
+        # Modify one of the values in the matrix
+        with connection.cursor() as cursor:
+            query = sql.SQL(
+                'UPDATE {0} SET {1} = true WHERE {2} = {3}').format(
+                    sql.Identifier(workflow.get_data_frame_table_name()),
+                    sql.Identifier('registered'),
+                    sql.Identifier('email'),
+                    sql.Placeholder())
+            cursor.execute(query, ['student01@bogus.com'])
+
+        # Execute the scheduler for the first time
+        tasks.execute_scheduled_actions_task(True)
+
+        # Event stil pending, with no values in exclude values
+        scheduled_item.refresh_from_db()
+        assert scheduled_item.status == ScheduledAction.STATUS_PENDING
+        assert scheduled_item.exclude_values == ['student01@bogus.com']
+
+        # Modify one of the values in the matrix
+        with connection.cursor() as cursor:
+            query = sql.SQL(
+                'UPDATE {0} SET {1} = true WHERE {2} = {3}').format(
+                    sql.Identifier(workflow.get_data_frame_table_name()),
+                    sql.Identifier('registered'),
+                    sql.Identifier('email'),
+                    sql.Placeholder())
+            cursor.execute(query, ['student02@bogus.com'])
+
+        # Execute the scheduler for the first time
+        tasks.execute_scheduled_actions_task(True)
+
+        # Event stil pending, with no values in exclude values
+        scheduled_item.refresh_from_db()
+        assert scheduled_item.status == ScheduledAction.STATUS_PENDING
+        assert scheduled_item.exclude_values == [
+            'student01@bogus.com',
+            'student02@bogus.com']
+
+        # Modify one of the values in the matrix
+        with connection.cursor() as cursor:
+            query = sql.SQL(
+                'UPDATE {0} SET {1} = true WHERE {2} = {3}').format(
+                    sql.Identifier(workflow.get_data_frame_table_name()),
+                    sql.Identifier('registered'),
+                    sql.Identifier('email'),
+                    sql.Placeholder())
+            cursor.execute(query, ['student03@bogus.com'])
+
+        # Execute the scheduler for the first time
+        tasks.execute_scheduled_actions_task(True)
+
+        # Event stil pending, with no values in exclude values
+        scheduled_item.refresh_from_db()
+        assert scheduled_item.status == ScheduledAction.STATUS_PENDING
+        assert scheduled_item.exclude_values == [
+            'student01@bogus.com',
+            'student02@bogus.com',
+            'student03@bogus.com']
+
+        # Execute the scheduler for the first time
+        tasks.execute_scheduled_actions_task(True)
+
+        # Event stil pending, with no values in exclude values
+        scheduled_item.refresh_from_db()
+        assert scheduled_item.status == ScheduledAction.STATUS_PENDING
+        assert scheduled_item.exclude_values == [
+            'student01@bogus.com',
+            'student02@bogus.com',
+            'student03@bogus.com']